pyproject.toml
[project]
name = "royal-blue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "pg8000>=1.31.2",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-testdox>=3.1.0",
    "pytest-cov>=6.1.1",
    "ruff==0.11.11",
    "moto>=5.1.5",
    "bandit>=1.8.3",
    "boto3>=1.38.23",
    "pandas>=2.2.3",
]

[tool.pytest.ini_options]
addopts = "-vvv -rP --testdox"
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"

[tool.ruff]
line-length = 88

[tool.ruff.lint]

[tool.ruff.format]
docstring-code-format = true
